WeatherFunction/requirements.txt
requests==2.25.1
azure-functions==1.7.0
